requirements.txt
Django==1.8
Pillow==2.8.0
